itinerary_group avanzando en los metodos del create

+++ am-core-web-service/src/main/java/org/am/core/web/dto/schedule/GroupDto.java
@@ -2,7 +2,7 @@ package org.am.core.web.dto.schedule;
 
 import java.util.List;
 
-public record GroupDto(Integer level,
+public record GroupDto(Short level,
                        String subjectName,
                        String subjectInitials,
                        String groupIdentifier,
